Update EntityFramework Project
