+++ angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFramework/EntityFramework/MyProjectDbContext.cs
@@ -1,0 +1,54 @@
+﻿using System.Data.Common;
+using System.Data.Entity;
+using Abp.Zero.EntityFramework;
+using MyDemo.MyProject.Authorization.Roles;
+using MyDemo.MyProject.Authorization.Users;
+using MyDemo.MyProject.MultiTenancy;
+
+namespace MyDemo.MyProject.EntityFramework
+{
+    public class MyProjectDbContext : AbpZeroDbContext<Tenant, Role, User, MyProjectDbContext>
+    {
+        //TODO: Define an IDbSet for your Entities...
+
+        /* NOTE: 
+         *   Setting "Default" to base class helps us when working migration commands on Package Manager Console.
+         *   But it may cause problems when working Migrate.exe of EF. If you will apply migrations on command line, do not
+         *   pass connection string name to base classes. ABP works either way.
+         */
+        public MyProjectDbContext()
+            : base("Default")
+        {
+
+        }
+
+        /* NOTE:
+         *   This constructor is used by ABP to pass connection string defined in AbpProjectNameDataModule.PreInitialize.
+         *   Notice that, actually you will not directly create an instance of MyProjectDbContext since ABP automatically handles it.
+         */
+        public MyProjectDbContext(string nameOrConnectionString)
+            : base(nameOrConnectionString)
+        {
+
+        }
+
+        //This constructor is used in tests
+        public MyProjectDbContext(DbConnection existingConnection)
+         : base(existingConnection, false)
+        {
+
+        }
+
+        public MyProjectDbContext(DbConnection existingConnection, bool contextOwnsConnection)
+         : base(existingConnection, contextOwnsConnection)
+        {
+
+        }
+
+        protected override void OnModelCreating(DbModelBuilder modelBuilder)
+        {
+            base.OnModelCreating(modelBuilder);
+
+        }
+    }
+}
