+++ angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFramework/EntityFramework/Seed/Host/DefaultEditionCreator.cs
@@ -1,0 +1,51 @@
+using System.Linq;
+using Abp.Application.Editions;
+using Abp.Application.Features;
+using MyDemo.MyProject.Editions;
+
+namespace MyDemo.MyProject.EntityFramework.Seed.Host
+{
+    public class DefaultEditionCreator
+    {
+        private readonly MyProjectDbContext _context;
+
+        public DefaultEditionCreator(MyProjectDbContext context)
+        {
+            _context = context;
+        }
+
+        public void Create()
+        {
+            CreateEditions();
+        }
+
+        private void CreateEditions()
+        {
+            var defaultEdition = _context.Editions.FirstOrDefault(e => e.Name == EditionManager.DefaultEditionName);
+            if (defaultEdition == null)
+            {
+                defaultEdition = new Edition { Name = EditionManager.DefaultEditionName, DisplayName = EditionManager.DefaultEditionName };
+                _context.Editions.Add(defaultEdition);
+                _context.SaveChanges();
+
+                /* Add desired features to the standard edition, if wanted... */
+            }
+        }
+
+        private void CreateFeatureIfNotExists(int editionId, string featureName, bool isEnabled)
+        {
+            if (_context.EditionFeatureSettings.Any(ef => ef.EditionId == editionId && ef.Name == featureName))
+            {
+                return;
+            }
+
+            _context.EditionFeatureSettings.Add(new EditionFeatureSetting
+            {
+                Name = featureName,
+                Value = isEnabled.ToString(),
+                EditionId = editionId
+            });
+            _context.SaveChanges();
+        }
+    }
+}
