+++ angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFramework/EntityFramework/Seed/Host/HostRoleAndUserCreator.cs
@@ -1,0 +1,97 @@
+using System.Linq;
+using Abp.Authorization;
+using Abp.Authorization.Roles;
+using Abp.Authorization.Users;
+using Abp.MultiTenancy;
+using MyDemo.MyProject.Authorization;
+using MyDemo.MyProject.Authorization.Roles;
+using MyDemo.MyProject.Authorization.Users;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.Extensions.Options;
+
+namespace MyDemo.MyProject.EntityFramework.Seed.Host
+{
+    public class HostRoleAndUserCreator
+    {
+        private readonly MyProjectDbContext _context;
+
+        public HostRoleAndUserCreator(MyProjectDbContext context)
+        {
+            _context = context;
+        }
+
+        public void Create()
+        {
+            CreateHostRoleAndUsers();
+        }
+
+        private void CreateHostRoleAndUsers()
+        {
+            // Admin role for host
+
+            var adminRoleForHost = _context.Roles.FirstOrDefault(r => r.TenantId == null && r.Name == StaticRoleNames.Host.Admin);
+            if (adminRoleForHost == null)
+            {
+                adminRoleForHost = _context.Roles.Add(new Role(null, StaticRoleNames.Host.Admin, StaticRoleNames.Host.Admin) { IsStatic = true, IsDefault = true });
+                _context.SaveChanges();
+            }
+
+            // Grant all permissions to admin role for host
+
+            var grantedPermissions = _context.Permissions
+                .OfType<RolePermissionSetting>()
+                .Where(p => p.TenantId == null && p.RoleId == adminRoleForHost.Id)
+                .Select(p => p.Name)
+                .ToList();
+
+            var permissions = PermissionFinder
+                .GetAllPermissions(new MyProjectAuthorizationProvider())
+                .Where(p => p.MultiTenancySides.HasFlag(MultiTenancySides.Host) &&
+                            !grantedPermissions.Contains(p.Name))
+                .ToList();
+
+            if (permissions.Any())
+            {
+                _context.Permissions.AddRange(
+                    permissions.Select(permission => new RolePermissionSetting
+                    {
+                        TenantId = null,
+                        Name = permission.Name,
+                        IsGranted = true,
+                        RoleId = adminRoleForHost.Id
+                    })
+                );
+                _context.SaveChanges();
+            }
+
+            // Admin user for host
+
+            var adminUserForHost = _context.Users.FirstOrDefault(u => u.TenantId == null && u.UserName == AbpUserBase.AdminUserName);
+            if (adminUserForHost == null)
+            {
+                var user = new User
+                {
+                    TenantId = null,
+                    UserName = AbpUserBase.AdminUserName,
+                    Name = "admin",
+                    Surname = "admin",
+                    EmailAddress = "admin@aspnetboilerplate.com",
+                    IsEmailConfirmed = true,
+                    IsActive = true
+                };
+
+                user.Password = new PasswordHasher<User>(new OptionsWrapper<PasswordHasherOptions>(new PasswordHasherOptions())).HashPassword(user, "123qwe");
+                user.SetNormalizedNames();
+
+                adminUserForHost = _context.Users.Add(user);
+                _context.SaveChanges();
+
+                // Assign Admin role to admin user
+                _context.UserRoles.Add(new UserRole(null, adminUserForHost.Id, adminRoleForHost.Id));
+                _context.SaveChanges();
+
+                _context.SaveChanges();
+            }
+        }
+    }
+}
