+++ angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFramework/EntityFramework/Seed/Host/InitialHostDbBuilder.cs
@@ -1,0 +1,22 @@
+﻿namespace MyDemo.MyProject.EntityFramework.Seed.Host
+{
+    public class InitialHostDbBuilder
+    {
+        private readonly MyProjectDbContext _context;
+
+        public InitialHostDbBuilder(MyProjectDbContext context)
+        {
+            _context = context;
+        }
+
+        public void Create()
+        {
+            new DefaultEditionCreator(_context).Create();
+            new DefaultLanguagesCreator(_context).Create();
+            new HostRoleAndUserCreator(_context).Create();
+            new DefaultSettingsCreator(_context).Create();
+
+            _context.SaveChanges();
+        }
+    }
+}
