+++ angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFramework/EntityFramework/Seed/Tenants/DefaultTenantBuilder.cs
@@ -1,0 +1,42 @@
+﻿using System.Linq;
+using Abp.MultiTenancy;
+using MyDemo.MyProject.Editions;
+using MyDemo.MyProject.MultiTenancy;
+
+namespace MyDemo.MyProject.EntityFramework.Seed.Tenants
+{
+    public class DefaultTenantBuilder
+    {
+        private readonly MyProjectDbContext _context;
+
+        public DefaultTenantBuilder(MyProjectDbContext context)
+        {
+            _context = context;
+        }
+
+        public void Create()
+        {
+            CreateDefaultTenant();
+        }
+
+        private void CreateDefaultTenant()
+        {
+            // Default tenant
+
+            var defaultTenant = _context.Tenants.FirstOrDefault(t => t.TenancyName == AbpTenantBase.DefaultTenantName);
+            if (defaultTenant == null)
+            {
+                defaultTenant = new Tenant(AbpTenantBase.DefaultTenantName, AbpTenantBase.DefaultTenantName);
+
+                var defaultEdition = _context.Editions.FirstOrDefault(e => e.Name == EditionManager.DefaultEditionName);
+                if (defaultEdition != null)
+                {
+                    defaultTenant.EditionId = defaultEdition.Id;
+                }
+
+                _context.Tenants.Add(defaultTenant);
+                _context.SaveChanges();
+            }
+        }
+    }
+}
