+++ angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFramework/EntityFramework/Seed/Tenants/TenantRoleAndUserBuilder.cs
@@ -1,0 +1,88 @@
+﻿using System.Linq;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.Extensions.Options;
+using Abp.Authorization;
+using Abp.Authorization.Roles;
+using Abp.Authorization.Users;
+using Abp.MultiTenancy;
+using MyDemo.MyProject.Authorization;
+using MyDemo.MyProject.Authorization.Roles;
+using MyDemo.MyProject.Authorization.Users;
+
+namespace MyDemo.MyProject.EntityFramework.Seed.Tenants
+{
+    public class TenantRoleAndUserBuilder
+    {
+        private readonly MyProjectDbContext _context;
+        private readonly int _tenantId;
+
+        public TenantRoleAndUserBuilder(MyProjectDbContext context, int tenantId)
+        {
+            _context = context;
+            _tenantId = tenantId;
+        }
+
+        public void Create()
+        {
+            CreateRolesAndUsers();
+        }
+
+        private void CreateRolesAndUsers()
+        {
+            // Admin role
+
+            var adminRole = _context.Roles.FirstOrDefault(r => r.TenantId == _tenantId && r.Name == StaticRoleNames.Tenants.Admin);
+            if (adminRole == null)
+            {
+                adminRole = _context.Roles.Add(new Role(_tenantId, StaticRoleNames.Tenants.Admin, StaticRoleNames.Tenants.Admin) { IsStatic = true });
+                _context.SaveChanges();
+            }
+
+            // Grant all permissions to admin role
+
+            var grantedPermissions = _context.Permissions
+                .OfType<RolePermissionSetting>()
+                .Where(p => p.TenantId == _tenantId && p.RoleId == adminRole.Id)
+                .Select(p => p.Name)
+                .ToList();
+
+            var permissions = PermissionFinder
+                .GetAllPermissions(new MyProjectAuthorizationProvider())
+                .Where(p => p.MultiTenancySides.HasFlag(MultiTenancySides.Tenant) &&
+                            !grantedPermissions.Contains(p.Name))
+                .ToList();
+
+            if (permissions.Any())
+            {
+                _context.Permissions.AddRange(
+                    permissions.Select(permission => new RolePermissionSetting
+                    {
+                        TenantId = _tenantId,
+                        Name = permission.Name,
+                        IsGranted = true,
+                        RoleId = adminRole.Id
+                    })
+                );
+                _context.SaveChanges();
+            }
+
+            // Admin user
+
+            var adminUser = _context.Users.FirstOrDefault(u => u.TenantId == _tenantId && u.UserName == AbpUserBase.AdminUserName);
+            if (adminUser == null)
+            {
+                adminUser = User.CreateTenantAdminUser(_tenantId, "admin@defaulttenant.com");
+                adminUser.Password = new PasswordHasher<User>(new OptionsWrapper<PasswordHasherOptions>(new PasswordHasherOptions())).HashPassword(adminUser, "123qwe");
+                adminUser.IsEmailConfirmed = true;
+                adminUser.IsActive = true;
+
+                _context.Users.Add(adminUser);
+                _context.SaveChanges();
+
+                // Assign Admin role to admin user
+                _context.UserRoles.Add(new UserRole(_tenantId, adminUser.Id, adminRole.Id));
+                _context.SaveChanges();
+            }
+        }
+    }
+}
