--- angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFrameworkCore/EntityFrameworkCore/AbpZeroDbMigrator.cs
@@ -1,21 +1,0 @@
-﻿using Abp.Domain.Uow;
-using Abp.EntityFrameworkCore;
-using Abp.MultiTenancy;
-using Abp.Zero.EntityFrameworkCore;
-
-namespace MyDemo.MyProject.EntityFrameworkCore
-{
-    public class AbpZeroDbMigrator : AbpZeroDbMigrator<MyProjectDbContext>
-    {
-        public AbpZeroDbMigrator(
-            IUnitOfWorkManager unitOfWorkManager,
-            IDbPerTenantConnectionStringResolver connectionStringResolver,
-            IDbContextResolver dbContextResolver)
-            : base(
-                unitOfWorkManager,
-                connectionStringResolver,
-                dbContextResolver)
-        {
-        }
-    }
-}
