--- angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFrameworkCore/EntityFrameworkCore/MyProjectDbContextConfigurer.cs
@@ -1,18 +1,0 @@
-using System.Data.Common;
-using Microsoft.EntityFrameworkCore;
-
-namespace MyDemo.MyProject.EntityFrameworkCore
-{
-    public static class MyProjectDbContextConfigurer
-    {
-        public static void Configure(DbContextOptionsBuilder<MyProjectDbContext> builder, string connectionString)
-        {
-            builder.UseSqlServer(connectionString);
-        }
-
-        public static void Configure(DbContextOptionsBuilder<MyProjectDbContext> builder, DbConnection connection)
-        {
-            builder.UseSqlServer(connection);
-        }
-    }
-}
