--- angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFrameworkCore/EntityFrameworkCore/MyProjectEntityFrameworkModule.cs
@@ -1,50 +1,0 @@
-﻿using Abp.EntityFrameworkCore.Configuration;
-using Abp.Modules;
-using Abp.Reflection.Extensions;
-using Abp.Zero.EntityFrameworkCore;
-using MyDemo.MyProject.EntityFrameworkCore.Seed;
-
-namespace MyDemo.MyProject.EntityFrameworkCore
-{
-    [DependsOn(
-        typeof(MyProjectCoreModule), 
-        typeof(AbpZeroCoreEntityFrameworkCoreModule))]
-    public class MyProjectEntityFrameworkModule : AbpModule
-    {
-        /* Used it tests to skip dbcontext registration, in order to use in-memory database of EF Core */
-        public bool SkipDbContextRegistration { get; set; }
-
-        public bool SkipDbSeed { get; set; }
-
-        public override void PreInitialize()
-        {
-            if (!SkipDbContextRegistration)
-            {
-                Configuration.Modules.AbpEfCore().AddDbContext<MyProjectDbContext>(options =>
-                {
-                    if (options.ExistingConnection != null)
-                    {
-                        MyProjectDbContextConfigurer.Configure(options.DbContextOptions, options.ExistingConnection);
-                    }
-                    else
-                    {
-                        MyProjectDbContextConfigurer.Configure(options.DbContextOptions, options.ConnectionString);
-                    }
-                });
-            }
-        }
-
-        public override void Initialize()
-        {
-            IocManager.RegisterAssemblyByConvention(typeof(MyProjectEntityFrameworkModule).GetAssembly());
-        }
-
-        public override void PostInitialize()
-        {
-            if (!SkipDbSeed)
-            {
-                SeedHelper.SeedHostDb(IocManager);
-            }
-        }
-    }
-}
