--- angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFrameworkCore/EntityFrameworkCore/Repositories/MyProjectRepositoryBase.cs
@@ -1,39 +1,0 @@
-﻿using Abp.Domain.Entities;
-using Abp.Domain.Repositories;
-using Abp.EntityFrameworkCore;
-using Abp.EntityFrameworkCore.Repositories;
-
-namespace MyDemo.MyProject.EntityFrameworkCore.Repositories
-{
-    /// <summary>
-    /// Base class for custom repositories of the application.
-    /// </summary>
-    /// <typeparam name="TEntity">Entity type</typeparam>
-    /// <typeparam name="TPrimaryKey">Primary key type of the entity</typeparam>
-    public abstract class MyProjectRepositoryBase<TEntity, TPrimaryKey> : EfCoreRepositoryBase<MyProjectDbContext, TEntity, TPrimaryKey>
-        where TEntity : class, IEntity<TPrimaryKey>
-    {
-        protected MyProjectRepositoryBase(IDbContextProvider<MyProjectDbContext> dbContextProvider)
-            : base(dbContextProvider)
-        {
-        }
-
-        // Add your common methods for all repositories
-    }
-
-    /// <summary>
-    /// Base class for custom repositories of the application.
-    /// This is a shortcut of <see cref="MyProjectRepositoryBase{TEntity,TPrimaryKey}"/> for <see cref="int"/> primary key.
-    /// </summary>
-    /// <typeparam name="TEntity">Entity type</typeparam>
-    public abstract class MyProjectRepositoryBase<TEntity> : MyProjectRepositoryBase<TEntity, int>, IRepository<TEntity>
-        where TEntity : class, IEntity<int>
-    {
-        protected MyProjectRepositoryBase(IDbContextProvider<MyProjectDbContext> dbContextProvider)
-            : base(dbContextProvider)
-        {
-        }
-
-        // Do not add any method here, add to the class above (since this inherits it)!!!
-    }
-}
