--- angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFrameworkCore/EntityFrameworkCore/Seed/Host/DefaultEditionCreator.cs
@@ -1,52 +1,0 @@
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
-using Abp.Application.Editions;
-using Abp.Application.Features;
-using MyDemo.MyProject.Editions;
-
-namespace MyDemo.MyProject.EntityFrameworkCore.Seed.Host
-{
-    public class DefaultEditionCreator
-    {
-        private readonly MyProjectDbContext _context;
-
-        public DefaultEditionCreator(MyProjectDbContext context)
-        {
-            _context = context;
-        }
-
-        public void Create()
-        {
-            CreateEditions();
-        }
-
-        private void CreateEditions()
-        {
-            var defaultEdition = _context.Editions.IgnoreQueryFilters().FirstOrDefault(e => e.Name == EditionManager.DefaultEditionName);
-            if (defaultEdition == null)
-            {
-                defaultEdition = new Edition { Name = EditionManager.DefaultEditionName, DisplayName = EditionManager.DefaultEditionName };
-                _context.Editions.Add(defaultEdition);
-                _context.SaveChanges();
-
-                /* Add desired features to the standard edition, if wanted... */
-            }
-        }
-
-        private void CreateFeatureIfNotExists(int editionId, string featureName, bool isEnabled)
-        {
-            if (_context.EditionFeatureSettings.IgnoreQueryFilters().Any(ef => ef.EditionId == editionId && ef.Name == featureName))
-            {
-                return;
-            }
-
-            _context.EditionFeatureSettings.Add(new EditionFeatureSetting
-            {
-                Name = featureName,
-                Value = isEnabled.ToString(),
-                EditionId = editionId
-            });
-            _context.SaveChanges();
-        }
-    }
-}
