--- angular/4.2.0/aspnet-core/src/MyDemo.MyProject.EntityFrameworkCore/EntityFrameworkCore/Seed/SeedHelper.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using System.Transactions;
-using Microsoft.EntityFrameworkCore;
-using Abp.Dependency;
-using Abp.Domain.Uow;
-using Abp.EntityFrameworkCore.Uow;
-using Abp.MultiTenancy;
-using MyDemo.MyProject.EntityFrameworkCore.Seed.Host;
-using MyDemo.MyProject.EntityFrameworkCore.Seed.Tenants;
-
-namespace MyDemo.MyProject.EntityFrameworkCore.Seed
-{
-    public static class SeedHelper
-    {
-        public static void SeedHostDb(IIocResolver iocResolver)
-        {
-            WithDbContext<MyProjectDbContext>(iocResolver, SeedHostDb);
-        }
-
-        public static void SeedHostDb(MyProjectDbContext context)
-        {
-            context.SuppressAutoSetTenantId = true;
-
-            // Host seed
-            new InitialHostDbBuilder(context).Create();
-
-            // Default tenant seed (in host database).
-            new DefaultTenantBuilder(context).Create();
-            new TenantRoleAndUserBuilder(context, 1).Create();
-        }
-
-        private static void WithDbContext<TDbContext>(IIocResolver iocResolver, Action<TDbContext> contextAction)
-            where TDbContext : DbContext
-        {
-            using (var uowManager = iocResolver.ResolveAsDisposable<IUnitOfWorkManager>())
-            {
-                using (var uow = uowManager.Object.Begin(TransactionScopeOption.Suppress))
-                {
-                    var context = uowManager.Object.Current.GetDbContext<TDbContext>(MultiTenancySides.Host);
-
-                    contextAction(context);
-
-                    uow.Complete();
-                }
-            }
-        }
-    }
-}
